Put split delimiter type mapping into detail and remove header guard
