+++ esl/macros.h
@@ -4,9 +4,6 @@
 
 #pragma once
 
-#ifndef ESL_MACROS_H_
-#define ESL_MACROS_H_
-
 #if !defined(__COUNTER__)
 #error "Requires __COUNTER__"
 #endif
@@ -14,5 +11,3 @@
 #define ESL_CONCAT_IMPL(a, b)  a##b
 #define ESL_CONCAT(a, b)       ESL_CONCAT_IMPL(a, b)
 #define ESL_ANONYMOUS_VAR(tag) ESL_CONCAT(ESL_CONCAT(ESL_CONCAT(tag, __LINE__), _), __COUNTER__)
-
-#endif // ESL_MACROS_H_
